Fix Bane of The Swine enchantment stacking and a refactor on a Charmed enchantment method
FIX: Fixed a bug where it was possible to stack the Bane of The Swing enchantment on a single weapon, making it very easy to one-hit a Piglin Brute(assuming the player stacked multiple Bane of The Swing V on a single weapon).
REFACTOR: The "canAccept" method of the Charmed enchantment was all a single line for some reason. I was trying to read it to see if another enchantment had the same stacking issue, and came across this. This has been refactored for easier readabilbity.