Deflection Enchantment
FEAT: Added the Deflection enchantment which gives a chance for projectiles to be deflected! A single level of Deflection isn't going to do much, but having full deflection 3 on all armor pieces will surely help a ton! Having 12 levels of Deflection will give an entity ~50% of deflecting an arrow!
REFACTOR: Corrected the Bane of Ender enchantment registry posistion