Explosive Enhancement & truly correct Bane of Ender registry pos.
FEAT: Added Explosive Enchantment. This enchantment applies to bows only, and sorta acts as an alternative to the crossbow's piercing enchantment(to a degree). Upon hitting an entity with this enchantment, an explosion gets created. The explosion's radius and power depends on the enchantment level.
REFACTOR: Truly corrected the Bane of Ender's registry position